Add application form for requesting to enrol on a degree or degree program
This commit introduces a simple application form at `/apply` that uses the
various FutureLearn Partners API endpoints to prefill form fields and display
the title of the degree or program the student is applying to join. This is all
facilitated by the `token` query string param that is provided when the student
is redirected from FutureLearn, which happens at the beginning of the
FutureLearn-first enrolment flow.