Add admin area for confirming enrolment requests
One of the final stages of the FutureLearn-first enrolment flow involves
confirming an enrolment request.

Before this step can happen the student must have an organisation membership –
to satisfy this step we're simply providing a text field to provide this on the
same form used to confirm the enrolment request.

We're also providing a similar shortcut for confirming degree enrolment
requests, where a single program can be specified (via code and run code text
fields) – the student will be enrolled on this program run after confirmation
happens. In reality these two 'shortcut' steps may happen separately.