Retrieve API details from the environment
To avoid committing sensitive credential info, this commit ensures we retrieve
these items from the environment instead.

We'll then be able to retrieve them globally within the app via
`Rails.configuration.x.api_config`.

There are multiple ways to ensure these variables are set, but the simplest is
to start a server like so:

  API_CONSUMER_KEY=your-consumer-key bundle exec rails server