Introduce services for interacting with the FutureLearn Partners API
To keep API request/response logic out of the controllers, this commit
introduces a `FutureLearnApi` module, which contains:

  - a class corresponding to each resource available via the Partners API, with
    methods for each endpoint

  - a `Connection` module for shared logic

As is the general theme with all the code in this simple Rails app, I am
favouring duplication over complex abstractions to make this code easier to
understand to non-Ruby/Rails developers, especially given it's unlikely
to change much.

Also note that we're explicitly raising an error if any responses fail due to
authentication errors (401) – this is the most likely error to occur and will
hopefully make debugging issues easier (rather than letting the JSON gem blow up
when trying to parse an empty response body).