Use SQLite database instead of PostgreSQL
Since this is intended as a sample application and is not expected to
run in production, our priority is simplicity over performance.

SQLite is also the database used for the Rails Getting Started Guide [1]
so this change should also make it easier for non-Ruby developers to get
the application running.