Store application details in local database
Previously we were storing the FutureLearn UUIDs for the learner and
the program run or degree they were applying for. In practice the
University is more likely to store the learner's name, contact details
and the University-assigned code for the program or degree. The only
FutureLearn issued UUID that the application portal needs to keep track
of is the enrolment request UUID which can be used to retrieve all
other details if necessary.