Introduce separate show and update actions for enrolment form
In preparation for demonstrating how a student might save a draft of
their application before submitting it, this commit replaces the `new`
and `create` actions on the enrolment requests controller with `show`
and `update`.

It also adds a separate `apply` action that receives the enrolment
request token and uses that to the appropriate the enrolment request
record in the local database, creating a new one based on the data
retrieved from the FutureLearn partners API if necessary.

Note that the `show` and `create` routes are currently vulnerable to
enumeration attacks. A production implementation would require
additional protection via an authentication token or similar.