Add 'Submit application' button to enrolment form
This button saves any changes on the form and then uses the patch
enrolment request endpoint on the FutureLearn partners API to transition
the enrolment request to the 'submitted' state. The return URL in the
response from the API is then used to redirect the user back to the
appropriate page in FutureLearn.