Add some basic integration tests to student recruitment API sample app
This commit adds some basic integration tests (Rails calls these
'system' tests) for the two main aspects the sample app is used for; receiving a
student lead notification, and viewing a list of student leads in the DB.

This will give us more confidence when we make changes. Writing these tests
stemmed from recently needing to update the app's dependencies and wanting
to have confidence that nothing had broken, without booting the app itself
and remembering how to interact with it manually.