Add some basic feature tests to FL-initiated enrolment sample app
This commit adds some basic feature tests (Rails calls these 'system' tests) for
the two main aspects the sample app is used for; simulating a user applying to
join a degree, and simulating the administration of enrolment requests.

These tests are (intentionally) far from exhaustive and only aim to alert us in
the case that we have broken something essential. Writing these tests stemmed
from recently needing to update the app's dependencies and wanting to have
confidence that nothing had broken, without booting the app itself and
remembering how to interact with it manually.

Note that I'm stubbing all external requests to the API, but only returning the
minimum data needed in each fake API response to get the tests to pass. I wanted
to avoid reproducing the Partners API to save time and keep the tests clearer:
the existing API documentation should be relied on for this.